Fix: build issue because of children not passed from UserProvider
