+++ nextjs-comic/package.json
@@ -20,12 +20,11 @@
     "dayjs": "^1.11.5",
     "jwt-decode": "^3.1.2",
     "lodash.debounce": "^4.0.8",
-    "next": "12.2.3",
+    "next": "^12.3.0",
     "normalize.css": "^8.0.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.35.0",
-    "react-router-dom": "^6.4.0",
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1"
   },
