+++ nextjs-comic/src/assets/scss/ComicDetail.module.scss
