+++ nextjs-comic/src/components/Carousel/Carousel.js
@@ -1,8 +1,8 @@
-import { useRef, forwardRef } from "react";
+import { forwardRef } from "react";
 import Slider from "react-slick";
 // Import css files
-import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
+import "slick-carousel/slick/slick.css";
 
 // eslint-disable-next-line react/display-name
 const Carousel = forwardRef((props, ref) => {
