+++ nextjs-comic/src/components/ComicList/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ComicList";
