--- nextjs-comic/src/layouts/components/ComicDetail/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ComicDetail";
