+++ nextjs-comic/src/services/comicService.js
@@ -1,5 +1,19 @@
-// import { makeRequest } from "~/utils/httpRequest";
+import { makeRequest } from "~/utils/httpRequest";
 
+export function getComics() {
+  return makeRequest(`/comics`, {
+    method: "GET",
+  });
+}
+
+export function getComicBySlug(slug) {
+  return makeRequest(`/comics/${slug}`, {
+    method: "GET",
+    params: {
+      test: true,
+    },
+  });
+}
 // export function getComics({ comicId, message, parentId }) {
 //   const {data, error, mutate} = useSWR(`comics`, fetcher)
